+++ pkg/detectors/checio/checio.go
@@ -2,10 +2,10 @@ package checio
 
 import (
 	"context"
-	// "log"
-	regexp "github.com/wasilibs/go-re2"
 	"net/http"
 	"strings"
+
+	regexp "github.com/wasilibs/go-re2"
 
 	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
